fix for hg support

+++ checkstyle-tester/launch.sh
@@ -11,26 +11,45 @@ while read line ; do
     [[ -z "$line" ]] && continue     # Skip empty lines
     
     REPO_NAME=`echo $line | cut -d '|' -f 1`
-    REPO_URL=` echo $line | cut -d '|' -f 2`
-    COMMIT_ID=`echo $line | cut -d '|' -f 3`
-    EXCLUDES=`echo $line | cut -d '|' -f 4`
+    REPO_TYPE=`echo $line | cut -d '|' -f 2`
+    REPO_URL=` echo $line | cut -d '|' -f 3`
+    COMMIT_ID=`echo $line | cut -d '|' -f 4`
+    EXCLUDES=` echo $line | cut -d '|' -f 5`
     
     EXCLUDES_ACCUM+=",$EXCLUDES"
 
     REPO_SOURCES_DIR=$SOURCES_DIR/$REPO_NAME
     
-    if [ ! -d "$REPO_SOURCES_DIR" ]; then
-        echo "Cloning repository '${REPO_NAME}' ..."
-        git clone $REPO_URL $REPO_SOURCES_DIR
-        echo -e "Cloning repository '$REPO_NAME' - completed\n"
-    fi
+    if [ "$REPO_TYPE" == "git" ]; then
+		if [ ! -d "$REPO_SOURCES_DIR" ]; then
+			echo "Cloning $REPO_TYPE repository '${REPO_NAME}' ..."
+			git clone $REPO_URL $REPO_SOURCES_DIR
+			echo -e "Cloning $REPO_TYPE repository '$REPO_NAME' - completed\n"
+			fi
 
-	if [ "$COMMIT_ID" != "" ]; then
-        echo "Reseting sources to commit '$COMMIT_ID'"
-        cd $REPO_SOURCES_DIR
-	    git reset --hard $COMMIT_ID
-        cd -
-	    fi
+		if [ "$COMMIT_ID" != "" ]; then
+			echo "Reseting $REPO_TYPE sources to commit '$COMMIT_ID'"
+			cd $REPO_SOURCES_DIR
+			git reset --hard $COMMIT_ID
+			cd -
+			fi
+    else
+        # esle is only hg (mercurial)
+		if [ ! -d "$REPO_SOURCES_DIR" ]; then
+			echo "Cloning $REPO_TYPE repository '${REPO_NAME}' ..."
+			hg clone $REPO_URL $REPO_SOURCES_DIR
+			echo -e "Cloning $REPO_TYPE repository '$REPO_NAME' - completed\n"
+			fi
+
+		if [ "$COMMIT_ID" != "" ]; then
+			echo "Reseting $REPO_TYPE sources to commit '$COMMIT_ID'"
+			cd $REPO_SOURCES_DIR
+			hg update --clean
+			cd -
+			fi 
+    
+    fi
+    
     echo -e "$REPO_NAME is synchronized\n"
 
 done < projects-to-test-on.properties
